Swifmeetup commit

+++ AbbreviationApp/AbbreviationApp/ListAcronymViewController.swift
@@ -26,7 +26,7 @@ class ListAcronymViewController: UIViewController,UITableViewDelegate,UITableVie
         let localDict : Dictionary = responseAcronyms?[indexPath.row] as! Dictionary<String,String>
         
         cellShortText  = localDict["short"]
-        cellLongText  = localDict["short"]
+        cellLongText  = localDict["long"]
         cell.textLabel?.text = cellShortText + " : " + cellLongText
 
         return cell
